+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Portable Abstractions and Patterns for Java Changelog
 
+## <a name="3.1.3"></a> 3.1.3 (2022-06-22)
+
+### Features
+- **data** getAsObject() added the ability to type the returned object
+- JsonConverter added ability to convert generics
+
 ## <a name="3.1.2"></a> 3.1.2 (2022-06-04)
 
 ### Features